choo, wrapper: exit on crash (#103)
wrapper: Add a `crash` parameter to the wrapper that toggles the
sending of an exit effect when a kernel is poked with a crud (generated
by the runtime on a crash when attempting to slam the poke arm of a
kernel).

choo: Update the choo moat to exit on crash.

Update extant NockApps with the new convention.